Removed redundant endpoints
