+++ server/categories-service/src/main/kotlin/xelagurd/socialdating/server/controller/UserCategoriesController.kt
@@ -1,21 +1,14 @@
 package xelagurd.socialdating.server.controller
 
-import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
-import jakarta.validation.Valid
 import xelagurd.socialdating.server.model.UserCategory
 import xelagurd.socialdating.server.model.additional.SimilarUser
 import xelagurd.socialdating.server.model.additional.UserWithSimilarity
-import xelagurd.socialdating.server.model.details.UserCategoryDetails
-import xelagurd.socialdating.server.security.AdminAccess
 import xelagurd.socialdating.server.service.UserCategoriesService
 
 @RestController
@@ -28,14 +21,6 @@ class UserCategoriesController(
     @GetMapping
     fun getUserCategories(@RequestParam userId: Int): List<UserCategory> {
         return userCategoriesService.getUserCategories(userId)
-    }
-
-    @Operation(security = [SecurityRequirement("bearerAuth")])
-    @AdminAccess
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    fun addUserCategory(@RequestBody @Valid userCategoryDetails: UserCategoryDetails): UserCategory {
-        return userCategoriesService.addUserCategory(userCategoryDetails)
     }
 
     @Operation(security = [SecurityRequirement("bearerAuth")])
