+++ server/categories-service/src/test/kotlin/xelagurd/socialdating/server/test/CategoriesMicroserviceTest.kt
@@ -13,10 +13,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.testcontainers.containers.PostgreSQLContainer
 import xelagurd.socialdating.server.FakeCategoriesData
-import xelagurd.socialdating.server.FakeCategoriesData.filterByUserId
-import xelagurd.socialdating.server.FakeCategoriesData.toServerAnswer
 import xelagurd.socialdating.server.model.Category
-import xelagurd.socialdating.server.model.UserCategory
 
 @ActiveProfiles("dev", "test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -24,26 +21,16 @@ import xelagurd.socialdating.server.model.UserCategory
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class CategoriesMicroserviceTest(@param:Autowired val restTemplate: TestRestTemplate) {
 
-    private val userId = 1
-
     private val categoriesDetails = FakeCategoriesData.categoriesDetails
     private val categories = FakeCategoriesData.categories.take(categoriesDetails.size)
-    private val userCategoriesDetails = FakeCategoriesData.userCategoriesDetails
-    private val userCategories = FakeCategoriesData.userCategories.take(userCategoriesDetails.size).toServerAnswer()
 
     init {
         addCategories()
-        addUserCategories()
     }
 
     @Test
     fun testGetCategories() {
         getCategories()
-    }
-
-    @Test
-    fun testGetUserCategories() {
-        getUserCategories()
     }
 
     private fun addCategories() {
@@ -58,18 +45,6 @@ class CategoriesMicroserviceTest(@param:Autowired val restTemplate: TestRestTemp
         }
     }
 
-    private fun addUserCategories() {
-        userCategoriesDetails.forEachIndexed { index, userCategoryDetails ->
-            val response = restTemplate.postForEntity(
-                "/categories/users",
-                userCategoryDetails,
-                UserCategory::class.java
-            )
-            assertEquals(HttpStatus.CREATED, response.statusCode)
-            assertEquals(userCategories[index], response.body!!)
-        }
-    }
-
     private fun getCategories() {
         val response = restTemplate.getForEntity(
             "/categories",
@@ -78,17 +53,6 @@ class CategoriesMicroserviceTest(@param:Autowired val restTemplate: TestRestTemp
         assertEquals(HttpStatus.OK, response.statusCode)
         assertEquals(categories.size, response.body!!.size)
         assertContentEquals(categories.toTypedArray(), response.body!!)
-    }
-
-    private fun getUserCategories() {
-        val expected = userCategories.filterByUserId(userId)
-        val response = restTemplate.getForEntity(
-            "/categories/users?userId=$userId",
-            Array<UserCategory>::class.java
-        )
-        assertEquals(HttpStatus.OK, response.statusCode)
-        assertEquals(expected.size, response.body!!.size)
-        assertContentEquals(expected.toTypedArray(), response.body!!)
     }
 
     companion object {
