+++ server/defining-themes-service/src/main/kotlin/xelagurd/socialdating/server/controller/UserDefiningThemesController.kt
@@ -1,19 +1,12 @@
 package xelagurd.socialdating.server.controller
 
-import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
-import jakarta.validation.Valid
 import xelagurd.socialdating.server.model.UserDefiningTheme
-import xelagurd.socialdating.server.model.details.UserDefiningThemeDetails
-import xelagurd.socialdating.server.security.AdminAccess
 import xelagurd.socialdating.server.service.UserDefiningThemesService
 
 @RestController
@@ -28,11 +21,4 @@ class UserDefiningThemesController(
         return userDefiningThemesService.getUserDefiningThemes(userId)
     }
 
-    @Operation(security = [SecurityRequirement("bearerAuth")])
-    @AdminAccess
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    fun addUserDefiningTheme(@RequestBody @Valid userDefiningThemeDetails: UserDefiningThemeDetails): UserDefiningTheme {
-        return userDefiningThemesService.addUserDefiningTheme(userDefiningThemeDetails)
-    }
 }
