+++ server/defining-themes-service/src/main/kotlin/xelagurd/socialdating/server/service/UserDefiningThemesService.kt
@@ -3,7 +3,6 @@ package xelagurd.socialdating.server.service
 import org.springframework.stereotype.Service
 import xelagurd.socialdating.server.exception.NoDataFoundException
 import xelagurd.socialdating.server.model.UserDefiningTheme
-import xelagurd.socialdating.server.model.details.UserDefiningThemeDetails
 import xelagurd.socialdating.server.repository.UserDefiningThemesRepository
 
 @Service
@@ -14,10 +13,6 @@ class UserDefiningThemesService(
     fun getUserDefiningThemes(userId: Int): List<UserDefiningTheme> {
         return userDefiningThemesRepository.findAllByUserId(userId).takeIf { it.isNotEmpty() }
             ?: throw NoDataFoundException("UserDefiningThemes didn't found for userId: $userId")
-    }
-
-    fun addUserDefiningTheme(userDefiningThemeDetails: UserDefiningThemeDetails): UserDefiningTheme {
-        return userDefiningThemesRepository.save(userDefiningThemeDetails.toUserDefiningTheme())
     }
 
     fun addUserDefiningTheme(userDefiningTheme: UserDefiningTheme): UserDefiningTheme {
