+++ server/statements-service/src/main/kotlin/xelagurd/socialdating/server/service/UserStatementsService.kt
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service
 import xelagurd.socialdating.server.exception.NoDataFoundException
 import xelagurd.socialdating.server.model.UserStatement
 import xelagurd.socialdating.server.model.additional.StatementReactionDetails
-import xelagurd.socialdating.server.model.details.UserStatementDetails
 import xelagurd.socialdating.server.repository.UserStatementsRepository
 
 @Service
@@ -19,17 +18,8 @@ class UserStatementsService(
             ?: throw NoDataFoundException("UserStatements didn't found for userId $userId and definingThemeIds $definingThemeIds")
     }
 
-    fun addUserStatement(userStatementDetails: UserStatementDetails): UserStatement {
-        return userStatementsRepository.save(userStatementDetails.toUserStatement())
-    }
-
     fun processStatementReaction(statementReactionDetails: StatementReactionDetails): UserStatement {
-        val userStatementDetails = UserStatementDetails(
-            reactionType = statementReactionDetails.reactionType,
-            userId = statementReactionDetails.userId,
-            statementId = statementReactionDetails.statementId
-        )
-        val userStatement = userStatementsRepository.save(userStatementDetails.toUserStatement())
+        val userStatement = userStatementsRepository.save(statementReactionDetails.toUserStatement())
 
         kafkaProducer.updateUserCategory(
             statementReactionDetails.toUserCategoryUpdateDetails()
